safari-600.7-branch/LayoutTests/svg/animations/animate-color-rgba-calcMode-discrete-expected.txt

SVG 1.1 dynamic animation tests

Test calcMode discrete with from-to animation on colors with alpha channel. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rgbColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rgbColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS rgbColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS parseAlphaFromColor() is 0.4
PASS rgbColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS rgbColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rgbColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS parseAlphaFromColor() is 0.6
PASS rgbColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS rgbColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rgbColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS parseAlphaFromColor() is 0.6
PASS rgbColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rgbColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS rgbColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS parseAlphaFromColor() is 0.8
PASS rgbColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rgbColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS rgbColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS parseAlphaFromColor() is 0.8
PASS rgbColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rgbColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS rgbColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS parseAlphaFromColor() is 0.4
PASS successfullyParsed is true

TEST COMPLETE
